+++ ch03/step_function.py
@@ -4,7 +4,7 @@ import matplotlib.pylab as plt
 
 
 def step_function(x):
-    return np.array(x > 0, dtype=np.int)
+    return np.array(x > 0, dtype=np.int32)
 
 X = np.arange(-5.0, 5.0, 0.1)
 Y = step_function(X)
